tests: convert passt tests to passt plugin
Only one passt (old API - non plugin) test is left.
